Add files via upload
I finished this a while ago but was trying to manually plot more of an elliptical path instead of a circular path and lost track of time. I didn't get very far but either I didn't do it correctly or keyframes are not the way to go about that. But it was a cool little project and I might add some more planetary objects at some point.